Extend SubGhzProtocolDecoder with ARF fields; add hash_data_long; fix sign compare
- types.h: add SubGhzGetHashDataLong typedef + get_hash_data_long and
  get_string_brief fields to SubGhzProtocolDecoder (NULL in stock protocols)
- decoder.h/c: add subghz_protocol_blocks_get_hash_data_long returning
  uint32_t for Keeloq's wider collision-resistant hash
- keeloq.h: change get_hash_data declaration to uint32_t to match impl
- keeloq.c: cast rolling_counter_mult result to int32_t before comparing
  to -0x7FFFFFFF sentinel (fixes -Werror=sign-compare)

+++ modified_files/keeloq.c
@@ -284,7 +284,7 @@ static bool subghz_protocol_keeloq_gen_data(
 
         if(keeloq_counter_mode == 0 || bypass) {
             // Check for OFEX (overflow experimental) mode
-            if(furi_hal_subghz_get_rolling_counter_mult() != -0x7FFFFFFF || bypass) {
+            if((int32_t)furi_hal_subghz_get_rolling_counter_mult() != -0x7FFFFFFF || bypass) {
                 bypass = false;
                 // standart counter mode. PULL data from subghz_block_generic_global variables
                 if(!subghz_block_generic_global_counter_override_get(&instance->generic.cnt)) {
